More Mask Sys Polish, New Mask
I polished up the mask system even more, so now it should be even smoother. I added comments where I added new stuff, so be sure to read if you're curious. I also added the fundamentals for a time slow mask, and I commented in the script my plans for it. Please be sure to read it if you're a fellow programmer.